requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
crewai==0.102.0
google-generativeai==0.8.4
langchain-google-genai==2.0.9
python-dotenv==1.0.1
pydantic==2.10.6
